+++ tools/schemacode/bidsschematools/validate_schema.py
